+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Show only valid actions for the current site.
 * Use GM_setClipboard to set the clipboard.
 * Add a setting: Trim Subtitle.
+* Skip trial books on Books site.
 
 # v0.5
 * Support Amazon CN.