+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # UPCOMING
+
+# v0.6
 * Show only valid actions for the current site.
 * Use GM_setClipboard to set the clipboard.
 * Add a setting: Trim Subtitle.